Full rewrite
* core/shell
* async via runaway
* plugins to implement shell

The plan had been incremental changes, but I wanted to rip out asyncio,
and that basically meant a full rewrite right there, so I did it.  In
the process I also experimented with core/shell/plugin architecture and
made use of the runaway project for async.

Had to implement my own plugin manager in the process, modeled off of
some of the literature & implementation of stevedore, but with a focus
on letting any plugin implement any shell function.  This manager still
needs more work... everything still needs more work, but from here on it
can be done more incrementally.

The present state is functional, the core is unit tested, and the whole
is E2E tested.